+++ packages/nextra-theme-docs/src/components/flexsearch.tsx
@@ -1,11 +1,12 @@
-import { useState, ReactElement, ReactNode, useCallback } from 'react'
+import type { ReactElement, ReactNode} from 'react';
+import { useState, useCallback } from 'react'
 import { useRouter } from 'next/router'
 import FlexSearch from 'flexsearch'
 import cn from 'clsx'
 import { Search } from './search'
 import { HighlightMatches } from './highlight-matches'
 import { DEFAULT_LOCALE } from '../constants'
-import { SearchResult } from '../types'
+import type { SearchResult } from '../types'
 
 type SectionIndex = FlexSearch.Document<
   {
@@ -153,6 +154,7 @@ export function Flexsearch({
 }): ReactElement {
   const { locale = DEFAULT_LOCALE, basePath } = useRouter()
   const [loading, setLoading] = useState(false)
+  const [error, setError] = useState(false)
   const [results, setResults] = useState<SearchResult[]>([])
   const [search, setSearch] = useState('')
 
@@ -251,7 +253,11 @@ export function Flexsearch({
     async (active: boolean) => {
       if (active && !indexes[locale]) {
         setLoading(true)
-        await loadIndexes(basePath, locale)
+        try {
+          await loadIndexes(basePath, locale)
+        } catch (e) {
+          setError(true)
+        }
         setLoading(false)
       }
     },
@@ -265,7 +271,11 @@ export function Flexsearch({
     }
     if (!indexes[locale]) {
       setLoading(true)
-      await loadIndexes(basePath, locale)
+      try {
+        await loadIndexes(basePath, locale)
+      } catch (e) {
+        setError(true)
+      }
       setLoading(false)
     }
     doSearch(value)
@@ -274,6 +284,7 @@ export function Flexsearch({
   return (
     <Search
       loading={loading}
+      error={error}
       value={search}
       onChange={handleChange}
       onActive={preload}