+++ packages/nextra-theme-docs/src/contexts/config.tsx
@@ -1,24 +1,23 @@
-import {
-  createContext,
-  ReactElement,
-  ReactNode,
-  useContext,
-  useState
-} from 'react'
-import { PageOpts } from 'nextra'
+import type { ReactElement, ReactNode } from 'react'
+import { createContext, useContext, useState } from 'react'
+import type { PageOpts, PageMapItem, FrontMatter } from 'nextra'
 import { ThemeProvider } from 'next-themes'
-import { Context } from '../types'
+import type { Context } from '../types'
+import type { DocsThemeConfig } from '../constants'
 import {
   DEEP_OBJECT_KEYS,
   DEFAULT_THEME,
-  DocsThemeConfig,
+  metaSchema,
   themeSchema
 } from '../constants'
 import { MenuProvider } from './menu'
 import type { ZodError } from 'zod'
 
-type Config = DocsThemeConfig &
-  Pick<PageOpts, 'flexsearch' | 'newNextLinkBehavior' | 'title' | 'frontMatter'>
+type Config<FrontMatterType = FrontMatter> = DocsThemeConfig &
+  Pick<
+    PageOpts<FrontMatterType>,
+    'flexsearch' | 'newNextLinkBehavior' | 'title' | 'frontMatter'
+  >
 
 const ConfigContext = createContext<Config>({
   title: '',
@@ -26,13 +25,47 @@ const ConfigContext = createContext<Config>({
   ...DEFAULT_THEME
 })
 
-export const useConfig = () => useContext(ConfigContext)
+export function useConfig<FrontMatterType = FrontMatter>() {
+  // @ts-expect-error TODO: fix Type 'Config<{ [key: string]: any; }>' is not assignable to type 'Config<FrontMatterType>'.
+  return useContext<Config<FrontMatterType>>(ConfigContext)
+}
 
 let theme: DocsThemeConfig
-let isValidated = process.env.NODE_ENV === 'production'
+let isValidated = false
 
-function lowerCaseFirstLetter(s: string) {
-  return s.charAt(0).toLowerCase() + s.slice(1)
+function normalizeZodMessage(error: unknown): string {
+  return (error as ZodError).issues
+    .flatMap(issue => {
+      const themePath =
+        issue.path.length > 0 && `Path: "${issue.path.join('.')}"`
+      const unionErrors =
+        'unionErrors' in issue ? issue.unionErrors.map(normalizeZodMessage) : []
+      return [
+        [issue.message, themePath].filter(Boolean).join('. '),
+        ...unionErrors
+      ]
+    })
+    .join('\n')
+}
+
+function validateMeta(pageMap: PageMapItem[]) {
+  for (const pageMapItem of pageMap) {
+    if (pageMapItem.kind === 'Meta') {
+      for (const [key, data] of Object.entries(pageMapItem.data)) {
+        try {
+          metaSchema.parse(data)
+        } catch (error) {
+          console.error(
+            `[nextra-theme-docs] Error validating _meta.json file for "${key}" property.\n\n${normalizeZodMessage(
+              error
+            )}`
+          )
+        }
+      }
+    } else if (pageMapItem.kind === 'Folder') {
+      validateMeta(pageMapItem.children)
+    }
+  }
 }
 
 export const ConfigProvider = ({
@@ -56,25 +89,17 @@ export const ConfigProvider = ({
       ])
     )
   }
-  if (!isValidated) {
+  if (process.env.NODE_ENV !== 'production' && !isValidated) {
     try {
-      theme = themeSchema.parse(theme, {
-        errorMap: () => ({ message: 'Invalid theme config' })
-      })
-    } catch (err) {
-      console.error('[nextra] Error validating the theme config')
+      theme = themeSchema.parse(theme)
+    } catch (error) {
       console.error(
-        (err as ZodError).issues
-          .map(
-            issue =>
-              '[nextra] Error in config `' +
-              issue.path.join('.') +
-              '`: ' +
-              lowerCaseFirstLetter(issue.message || '')
-          )
-          .join('\n')
+        `[nextra-theme-docs] Error validating theme config file.\n\n${normalizeZodMessage(
+          error
+        )}`
       )
     }
+    validateMeta(pageOpts.pageMap)
     isValidated = true
   }
   const extendedConfig: Config = {