+++ packages/nextra-theme-docs/src/types.ts
@@ -1,20 +1,7 @@
 /* eslint typescript-sort-keys/interface: error */
-import { ReactNode } from 'react'
-import { PageOpts } from 'nextra'
-import { DocsThemeConfig } from './constants'
-
-export type PageTheme = {
-  breadcrumb: boolean
-  collapsed: boolean
-  footer: boolean
-  layout: 'default' | 'full' | 'raw'
-  navbar: boolean
-  pagination: boolean
-  sidebar: boolean
-  timestamp: boolean
-  toc: boolean
-  typesetting: 'default' | 'article'
-}
+import type { ReactNode } from 'react'
+import type { PageOpts } from 'nextra'
+import type { DocsThemeConfig } from './constants'
 
 export type Context = {
   pageOpts: PageOpts