+++ packages/nextra/src/setup-page.ts
@@ -6,13 +6,97 @@
 import type { FC } from 'react'
 import type {
   DynamicMetaDescriptor,
+  Folder,
   NextraInternalGlobal,
   PageOpts,
-  ThemeConfig
+  ThemeConfig,
+  PageMapItem,
+  DynamicMetaJsonFile,
+  DynamicMeta,
+  DynamicMetaItem,
+  DynamicFolder
 } from './types'
+import { normalizePageRoute, pageTitleFromFilename } from './utils'
 
 import get from 'lodash.get'
 import { NEXTRA_INTERNAL } from './constants'
+
+function isFolder(value: DynamicMetaItem): value is DynamicFolder {
+  return !!value && typeof value === 'object' && value.type === 'folder'
+}
+
+function normalizeMetaData(obj: DynamicMeta): DynamicMeta {
+  return Object.fromEntries(
+    Object.entries(obj).map(([key, value]) => {
+      if (isFolder(value)) {
+        const keyWithoutSlash = key.replace('/', '')
+        return [
+          keyWithoutSlash,
+          value.title || pageTitleFromFilename(keyWithoutSlash)
+        ]
+      }
+      return [key, value || pageTitleFromFilename(key)]
+    })
+  )
+}
+
+export function collectCatchAllRoutes(
+  parent: Folder<any>,
+  meta: DynamicMetaJsonFile,
+  isRootFolder = true
+): void {
+  if (isRootFolder) {
+    collectCatchAllRoutes(
+      parent,
+      {
+        kind: 'Meta',
+        data: meta.data,
+        locale: meta.locale
+      },
+      false
+    )
+    meta.data = normalizeMetaData(meta.data)
+    return
+  }
+  for (const [key, value] of Object.entries(meta.data)) {
+    if (!isFolder(value)) {
+      parent.children.push({
+        kind: 'MdxPage',
+        ...(meta.locale && { locale: meta.locale }),
+        name: key,
+        title: value || pageTitleFromFilename(key),
+        route: normalizePageRoute(parent.route, key)
+      })
+      continue
+    }
+    const routeWithoutSlashes = key.replace('/', '')
+    const newParent: Folder = {
+      kind: 'Folder',
+      name: routeWithoutSlashes,
+      route: `${parent.route}/${routeWithoutSlashes}`,
+      children: [
+        {
+          kind: 'Meta',
+          ...(meta.locale && { locale: meta.locale }),
+          data: normalizeMetaData(value.items)
+        }
+      ]
+    }
+
+    parent.children.push(newParent)
+    collectCatchAllRoutes(
+      newParent,
+      {
+        kind: 'Meta',
+        data: value.items,
+        ...(meta.locale && { locale: meta.locale })
+      },
+      false
+    )
+  }
+}
+
+let cachedResolvedPageMap: PageMapItem[]
 
 export function setupNextraPage({
   pageNextRoute,
@@ -35,31 +119,30 @@ export function setupNextraPage({
 }) {
   if (typeof window === 'undefined') {
     globalThis.__nextra_resolvePageMap = async () => {
-      const clonedPageMap = JSON.parse(JSON.stringify(pageOpts.pageMap))
+      if (process.env.NODE_ENV === 'production' && cachedResolvedPageMap) {
+        return cachedResolvedPageMap
+      }
+      const clonedPageMap: PageMapItem[] = JSON.parse(
+        JSON.stringify(pageOpts.pageMap)
+      )
+
       await Promise.all(
         dynamicMetaModules.map(
           async ([importMod, { metaObjectKeyPath, metaParentKeyPath }]) => {
             const mod = await importMod
             const metaData = await mod.default()
-            const meta = get(clonedPageMap, metaObjectKeyPath)
+            const meta: DynamicMetaJsonFile = get(
+              clonedPageMap,
+              metaObjectKeyPath
+            )
             meta.data = metaData
 
-            const parentRoute =
-              get(clonedPageMap, metaParentKeyPath.replace(/\.children$/, ''))
-                .route || ''
-
-            for (const key of Object.keys(metaData)) {
-              get(clonedPageMap, metaParentKeyPath).push({
-                kind: 'MdxPage',
-                locale: meta.locale,
-                name: key.split('/').pop(),
-                route: parentRoute + '/' + key
-              })
-            }
+            const parent: Folder = get(clonedPageMap, metaParentKeyPath)
+            collectCatchAllRoutes(parent, meta)
           }
         )
       )
-      return clonedPageMap
+      return (cachedResolvedPageMap = clonedPageMap)
     }
   }