+++ packages/nextra/src/ssg.tsx
@@ -1,6 +1,7 @@
 import { createContext, useContext } from 'react'
 import { MDXRemote } from 'next-mdx-remote'
 
+import type { Components } from './mdx';
 import { useMDXComponents } from './mdx'
 
 export const SSGContext = createContext<any>(false)
@@ -11,8 +12,13 @@ export const useSSG = (key = 'ssg') => useContext(SSGContext)?.[key]
 export const DataContext = SSGContext
 export const useData = useSSG
 
-export function RemoteContent() {
+export function RemoteContent({
+  components: dynamicComponents
+}: {
+  components: Components
+}) {
   const dynamicContext = useSSG('__nextra_dynamic_mdx')
+
   if (!dynamicContext) {
     throw new Error(
       'RemoteContent must be used together with the `buildDynamicMDX` API'
@@ -21,5 +27,10 @@ export function RemoteContent() {
 
   const components = useMDXComponents()
 
-  return <MDXRemote compiledSource={dynamicContext} components={components} />
+  return (
+    <MDXRemote
+      compiledSource={dynamicContext}
+      components={{ ...components, ...dynamicComponents }}
+    />
+  )
 }