+++ packages/nextra/src/utils.ts
@@ -1,7 +1,8 @@
 import path from 'node:path'
 import title from 'title'
+import slash from 'slash'
 import { LOCALE_REGEX } from './constants'
-import { Folder, MdxFile, Meta } from './types'
+import type { Folder, MdxFile, Meta } from './types'
 
 export function parseFileName(filePath: string): {
   name: string
@@ -28,6 +29,10 @@ export function normalizeMeta(meta: Meta): Exclude<Meta, string> {
   return typeof meta === 'string' ? { title: meta } : meta
 }
 
+export function normalizePageRoute(parentRoute: string, route: string): string {
+  return slash(path.join(parentRoute, route.replace(/^index$/, '')))
+}
+
 export function pageTitleFromFilename(fileName: string) {
   return title(fileName.replace(/[-_]/g, ' '))
 }
@@ -39,6 +44,9 @@ export function sortPages(
   )[],
   locale?: string
 ): [string, string][] {
+  if (locale === '') {
+    locale = undefined
+  }
   return pages
     .filter(item => item.kind === 'Folder' || item.locale === locale)
     .map(item => ({